Refactor to DRY up and consolidate
- colocate RowRemovalOperation tests by merging CacheTests
- DRY up checking code
- improve naming

Refactor to reduce redundancy & improve naming
